[refactor] JAXB Code should be generated
