+++ exist-core/pom.xml
@@ -789,6 +789,7 @@
                                 <include>src/test/resources-filtered/conf.xml</include>
                                 <include>src/test/resources/log4j2.xml</include>
                                 <include>src/test/resources/standalone-webapp/WEB-INF/web.xml</include>
+                                <include>src/main/xjb/rest-api.xjb</include>
                                 <include>src/test/xquery/tail-recursion.xml</include>
                                 <include>src/test/xquery/maps/maps.xqm</include>
                                 <include>src/test/xquery/numbers/format-numbers.xql</include>
@@ -1370,6 +1371,7 @@
                                 <exclude>src/test/resources-filtered/conf.xml</exclude>
                                 <exclude>src/test/resources/log4j2.xml</exclude>
                                 <exclude>src/test/resources/standalone-webapp/WEB-INF/web.xml</exclude>
+                                <exclude>src/main/xjb/rest-api.xjb</exclude>
                                 <exclude>src/test/xquery/binary-value.xqm</exclude>
                                 <exclude>src/test/xquery/instance-of.xqm</exclude>
                                 <exclude>src/test/xquery/operator-mapping.xqm</exclude>
@@ -2350,6 +2352,24 @@ The BaseX Team. The original license statement is also included below.]]></pream
                             <findRegex>new Integer</findRegex>
                             <replaceValue>Integer.valueOf</replaceValue>
                         </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+
+            <plugin>
+                <groupId>org.jvnet.jaxb</groupId>
+                <artifactId>jaxb-maven-plugin</artifactId>
+                <version>${jaxb.impl.version}</version>
+                <configuration>
+                    <bindingDirectory>src/main/xjb</bindingDirectory>
+                    <schemaDirectory>src/main/xsd</schemaDirectory>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>generate-jaxb-objects</id>
+                        <goals>
+                            <goal>generate</goal>
+                        </goals>
                     </execution>
                 </executions>
             </plugin>
