--- exist-core/src/main/java/org/exist/http/jaxb/ObjectFactory.java
@@ -1,158 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v3.0.0 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.12.01 at 10:16:19 AM CET 
-//
-
-/*
- * eXist-db Open Source Native XML Database
- * Copyright (C) 2001 The eXist-db Authors
- *
- * info@exist-db.org
- * http://www.exist-db.org
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-package org.exist.http.jaxb;
-
-import jakarta.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the org.exist.http.jaxb package. 
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: org.exist.http.jaxb
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link Query }
-     *
-     * @return the Query.
-     */
-    public Query createQuery() {
-        return new Query();
-    }
-
-    /**
-     * Create an instance of {@link Result }
-     *
-     * @return the Result.
-     */
-    public Result createResult() {
-        return new Result();
-    }
-
-    /**
-     * Create an instance of {@link Sequence }
-     *
-     * @return the Sequence.
-     */
-    public Sequence createSequence() {
-        return new Sequence();
-    }
-
-    /**
-     * Create an instance of {@link Query.Properties }
-     *
-     * @return the Query.Properties.
-     */
-    public Query.Properties createQueryProperties() {
-        return new Query.Properties();
-    }
-
-    /**
-     * Create an instance of {@link Query.Variables }
-     *
-     * @return the Query.Variables.
-     */
-    public Query.Variables createQueryVariables() {
-        return new Query.Variables();
-    }
-
-    /**
-     * Create an instance of {@link Query.Variables.Variable }
-     *
-     * @return the Query.Variable.
-     */
-    public Query.Variables.Variable createQueryVariablesVariable() {
-        return new Query.Variables.Variable();
-    }
-
-    /**
-     * Create an instance of {@link Collection }
-     *
-     * @return the Collection.
-     */
-    public Collection createCollection() {
-        return new Collection();
-    }
-
-    /**
-     * Create an instance of {@link Result.Value }
-     *
-     * @return the Result.Value.
-     */
-    public Result.Value createResultValue() {
-        return new Result.Value();
-    }
-
-    /**
-     * Create an instance of {@link Sequence.Value }
-     *
-     * @return the Sequence.Value.
-     */
-    public Sequence.Value createSequenceValue() {
-        return new Sequence.Value();
-    }
-
-    /**
-     * Create an instance of {@link Query.Properties.Property }
-     *
-     * @return The Query.Property.
-     */
-    public Query.Properties.Property createQueryPropertiesProperty() {
-        return new Query.Properties.Property();
-    }
-
-    /**
-     * Create an instance of {@link Query.Variables.Variable.Qname }
-     *
-     * @return the Query.Variable.Qname.
-     */
-    public Query.Variables.Variable.Qname createQueryVariablesVariableQname() {
-        return new Query.Variables.Variable.Qname();
-    }
-
-}
