--- exist-core/src/main/java/org/exist/http/jaxb/Query.java
@@ -1,883 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v3.0.0 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.12.01 at 10:16:19 AM CET 
-//
-
-/*
- * eXist-db Open Source Native XML Database
- * Copyright (C) 2001 The eXist-db Authors
- *
- * info@exist-db.org
- * http://www.exist-db.org
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-package org.exist.http.jaxb;
-
-import jakarta.xml.bind.annotation.*;
-import jakarta.xml.bind.annotation.adapters.CollapsedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-
-/**
- * <p>Java class for anonymous complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType&gt;
- *   &lt;complexContent&gt;
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *       &lt;sequence&gt;
- *         &lt;element name="text" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
- *         &lt;element name="variables" minOccurs="0"&gt;
- *           &lt;complexType&gt;
- *             &lt;complexContent&gt;
- *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *                 &lt;sequence&gt;
- *                   &lt;element name="variable"&gt;
- *                     &lt;complexType&gt;
- *                       &lt;complexContent&gt;
- *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *                           &lt;sequence&gt;
- *                             &lt;element name="qname"&gt;
- *                               &lt;complexType&gt;
- *                                 &lt;complexContent&gt;
- *                                   &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *                                     &lt;sequence&gt;
- *                                       &lt;element name="localname" type="{http://www.w3.org/2001/XMLSchema}NCName"/&gt;
- *                                       &lt;element name="namespace" type="{http://www.w3.org/2001/XMLSchema}anyURI"/&gt;
- *                                       &lt;element name="prefix" type="{http://www.w3.org/2001/XMLSchema}NCName" minOccurs="0"/&gt;
- *                                       &lt;element ref="{http://exist-db.org/xquery/types/serialized}sequence"/&gt;
- *                                     &lt;/sequence&gt;
- *                                   &lt;/restriction&gt;
- *                                 &lt;/complexContent&gt;
- *                               &lt;/complexType&gt;
- *                             &lt;/element&gt;
- *                           &lt;/sequence&gt;
- *                         &lt;/restriction&gt;
- *                       &lt;/complexContent&gt;
- *                     &lt;/complexType&gt;
- *                   &lt;/element&gt;
- *                 &lt;/sequence&gt;
- *               &lt;/restriction&gt;
- *             &lt;/complexContent&gt;
- *           &lt;/complexType&gt;
- *         &lt;/element&gt;
- *         &lt;element name="properties" minOccurs="0"&gt;
- *           &lt;complexType&gt;
- *             &lt;complexContent&gt;
- *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *                 &lt;sequence&gt;
- *                   &lt;element name="property" maxOccurs="unbounded"&gt;
- *                     &lt;complexType&gt;
- *                       &lt;complexContent&gt;
- *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *                           &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
- *                           &lt;attribute name="value" use="required" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
- *                         &lt;/restriction&gt;
- *                       &lt;/complexContent&gt;
- *                     &lt;/complexType&gt;
- *                   &lt;/element&gt;
- *                 &lt;/sequence&gt;
- *               &lt;/restriction&gt;
- *             &lt;/complexContent&gt;
- *           &lt;/complexType&gt;
- *         &lt;/element&gt;
- *       &lt;/sequence&gt;
- *       &lt;attribute name="start" type="{http://www.w3.org/2001/XMLSchema}int" /&gt;
- *       &lt;attribute name="max" type="{http://www.w3.org/2001/XMLSchema}int" /&gt;
- *       &lt;attribute name="enclose" type="{http://exist.sourceforge.net/NS/exist}yesNo" /&gt;
- *       &lt;attribute name="wrap" type="{http://exist.sourceforge.net/NS/exist}yesNo" /&gt;
- *       &lt;attribute name="method" type="{http://www.w3.org/2001/XMLSchema}NCName" /&gt;
- *       &lt;attribute name="typed" type="{http://exist.sourceforge.net/NS/exist}yesNo" /&gt;
- *       &lt;attribute name="mime" type="{http://www.w3.org/2001/XMLSchema}NCName" /&gt;
- *       &lt;attribute name="cache" type="{http://exist.sourceforge.net/NS/exist}yesNo" /&gt;
- *       &lt;attribute name="session" type="{http://www.w3.org/2001/XMLSchema}NCName" /&gt;
- *     &lt;/restriction&gt;
- *   &lt;/complexContent&gt;
- * &lt;/complexType&gt;
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = {
-    "text",
-    "variables",
-    "properties"
-})
-@XmlRootElement(name = "query")
-public class Query {
-
-    @XmlElement(required = true)
-    protected String text;
-    protected Query.Variables variables;
-    protected Query.Properties properties;
-    @XmlAttribute(name = "start")
-    protected Integer start;
-    @XmlAttribute(name = "max")
-    protected Integer max;
-    @XmlAttribute(name = "enclose")
-    protected YesNo enclose;
-    @XmlAttribute(name = "wrap")
-    protected YesNo wrap;
-    @XmlAttribute(name = "method")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    @XmlSchemaType(name = "NCName")
-    protected String method;
-    @XmlAttribute(name = "typed")
-    protected YesNo typed;
-    @XmlAttribute(name = "mime")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    @XmlSchemaType(name = "NCName")
-    protected String mime;
-    @XmlAttribute(name = "cache")
-    protected YesNo cache;
-    @XmlAttribute(name = "session")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    @XmlSchemaType(name = "NCName")
-    protected String session;
-
-    /**
-     * Gets the value of the text property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getText() {
-        return text;
-    }
-
-    /**
-     * Sets the value of the text property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setText(String value) {
-        this.text = value;
-    }
-
-    /**
-     * Gets the value of the variables property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Query.Variables }
-     *     
-     */
-    public Query.Variables getVariables() {
-        return variables;
-    }
-
-    /**
-     * Sets the value of the variables property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Query.Variables }
-     *     
-     */
-    public void setVariables(Query.Variables value) {
-        this.variables = value;
-    }
-
-    /**
-     * Gets the value of the properties property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Query.Properties }
-     *     
-     */
-    public Query.Properties getProperties() {
-        return properties;
-    }
-
-    /**
-     * Sets the value of the properties property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Query.Properties }
-     *     
-     */
-    public void setProperties(Query.Properties value) {
-        this.properties = value;
-    }
-
-    /**
-     * Gets the value of the start property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getStart() {
-        return start;
-    }
-
-    /**
-     * Sets the value of the start property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setStart(Integer value) {
-        this.start = value;
-    }
-
-    /**
-     * Gets the value of the max property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getMax() {
-        return max;
-    }
-
-    /**
-     * Sets the value of the max property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setMax(Integer value) {
-        this.max = value;
-    }
-
-    /**
-     * Gets the value of the enclose property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link YesNo }
-     *     
-     */
-    public YesNo getEnclose() {
-        return enclose;
-    }
-
-    /**
-     * Sets the value of the enclose property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link YesNo }
-     *     
-     */
-    public void setEnclose(YesNo value) {
-        this.enclose = value;
-    }
-
-    /**
-     * Gets the value of the wrap property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link YesNo }
-     *     
-     */
-    public YesNo getWrap() {
-        return wrap;
-    }
-
-    /**
-     * Sets the value of the wrap property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link YesNo }
-     *     
-     */
-    public void setWrap(YesNo value) {
-        this.wrap = value;
-    }
-
-    /**
-     * Gets the value of the method property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getMethod() {
-        return method;
-    }
-
-    /**
-     * Sets the value of the method property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setMethod(String value) {
-        this.method = value;
-    }
-
-    /**
-     * Gets the value of the typed property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link YesNo }
-     *     
-     */
-    public YesNo getTyped() {
-        return typed;
-    }
-
-    /**
-     * Sets the value of the typed property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link YesNo }
-     *     
-     */
-    public void setTyped(YesNo value) {
-        this.typed = value;
-    }
-
-    /**
-     * Gets the value of the mime property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getMime() {
-        return mime;
-    }
-
-    /**
-     * Sets the value of the mime property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setMime(String value) {
-        this.mime = value;
-    }
-
-    /**
-     * Gets the value of the cache property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link YesNo }
-     *     
-     */
-    public YesNo getCache() {
-        return cache;
-    }
-
-    /**
-     * Sets the value of the cache property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link YesNo }
-     *     
-     */
-    public void setCache(YesNo value) {
-        this.cache = value;
-    }
-
-    /**
-     * Gets the value of the session property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getSession() {
-        return session;
-    }
-
-    /**
-     * Sets the value of the session property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setSession(String value) {
-        this.session = value;
-    }
-
-
-    /**
-     * <p>Java class for anonymous complex type.
-     * 
-     * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
-     * <pre>
-     * &lt;complexType&gt;
-     *   &lt;complexContent&gt;
-     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-     *       &lt;sequence&gt;
-     *         &lt;element name="property" maxOccurs="unbounded"&gt;
-     *           &lt;complexType&gt;
-     *             &lt;complexContent&gt;
-     *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-     *                 &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
-     *                 &lt;attribute name="value" use="required" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
-     *               &lt;/restriction&gt;
-     *             &lt;/complexContent&gt;
-     *           &lt;/complexType&gt;
-     *         &lt;/element&gt;
-     *       &lt;/sequence&gt;
-     *     &lt;/restriction&gt;
-     *   &lt;/complexContent&gt;
-     * &lt;/complexType&gt;
-     * </pre>
-     * 
-     * 
-     */
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {
-        "property"
-    })
-    public static class Properties {
-
-        @XmlElement(required = true)
-        protected List<Query.Properties.Property> property;
-
-        /**
-         * Gets the value of the property property.
-         * 
-         * <p>
-         * This accessor method returns a reference to the live list,
-         * not a snapshot. Therefore any modification you make to the
-         * returned list will be present inside the Jakarta XML Binding object.
-         * This is why there is not a <CODE>set</CODE> method for the property property.
-         * 
-         * <p>
-         * For example, to add a new item, do as follows:
-         * <pre>
-         *    getProperty().add(newItem);
-         * </pre>
-         * 
-         * 
-         * <p>
-         * Objects of the following type(s) are allowed in the list
-         * {@link Query.Properties.Property }
-         * 
-         * @return the properties.
-         */
-        public List<Query.Properties.Property> getProperty() {
-            if (property == null) {
-                property = new ArrayList<Query.Properties.Property>();
-            }
-            return this.property;
-        }
-
-
-        /**
-         * <p>Java class for anonymous complex type.
-         * 
-         * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
-         * <pre>
-         * &lt;complexType&gt;
-         *   &lt;complexContent&gt;
-         *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-         *       &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
-         *       &lt;attribute name="value" use="required" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
-         *     &lt;/restriction&gt;
-         *   &lt;/complexContent&gt;
-         * &lt;/complexType&gt;
-         * </pre>
-         * 
-         * 
-         */
-        @XmlAccessorType(XmlAccessType.FIELD)
-        @XmlType(name = "")
-        public static class Property {
-
-            @XmlAttribute(name = "name", required = true)
-            protected String name;
-            @XmlAttribute(name = "value", required = true)
-            protected String value;
-
-            /**
-             * Gets the value of the name property.
-             * 
-             * @return
-             *     possible object is
-             *     {@link String }
-             *     
-             */
-            public String getName() {
-                return name;
-            }
-
-            /**
-             * Sets the value of the name property.
-             * 
-             * @param value
-             *     allowed object is
-             *     {@link String }
-             *     
-             */
-            public void setName(String value) {
-                this.name = value;
-            }
-
-            /**
-             * Gets the value of the value property.
-             * 
-             * @return
-             *     possible object is
-             *     {@link String }
-             *     
-             */
-            public String getValue() {
-                return value;
-            }
-
-            /**
-             * Sets the value of the value property.
-             * 
-             * @param value
-             *     allowed object is
-             *     {@link String }
-             *     
-             */
-            public void setValue(String value) {
-                this.value = value;
-            }
-
-        }
-
-    }
-
-
-    /**
-     * <p>Java class for anonymous complex type.
-     * 
-     * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
-     * <pre>
-     * &lt;complexType&gt;
-     *   &lt;complexContent&gt;
-     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-     *       &lt;sequence&gt;
-     *         &lt;element name="variable"&gt;
-     *           &lt;complexType&gt;
-     *             &lt;complexContent&gt;
-     *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-     *                 &lt;sequence&gt;
-     *                   &lt;element name="qname"&gt;
-     *                     &lt;complexType&gt;
-     *                       &lt;complexContent&gt;
-     *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-     *                           &lt;sequence&gt;
-     *                             &lt;element name="localname" type="{http://www.w3.org/2001/XMLSchema}NCName"/&gt;
-     *                             &lt;element name="namespace" type="{http://www.w3.org/2001/XMLSchema}anyURI"/&gt;
-     *                             &lt;element name="prefix" type="{http://www.w3.org/2001/XMLSchema}NCName" minOccurs="0"/&gt;
-     *                             &lt;element ref="{http://exist-db.org/xquery/types/serialized}sequence"/&gt;
-     *                           &lt;/sequence&gt;
-     *                         &lt;/restriction&gt;
-     *                       &lt;/complexContent&gt;
-     *                     &lt;/complexType&gt;
-     *                   &lt;/element&gt;
-     *                 &lt;/sequence&gt;
-     *               &lt;/restriction&gt;
-     *             &lt;/complexContent&gt;
-     *           &lt;/complexType&gt;
-     *         &lt;/element&gt;
-     *       &lt;/sequence&gt;
-     *     &lt;/restriction&gt;
-     *   &lt;/complexContent&gt;
-     * &lt;/complexType&gt;
-     * </pre>
-     * 
-     * 
-     */
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {
-        "variable"
-    })
-    public static class Variables {
-
-        @XmlElement(required = true)
-        protected Query.Variables.Variable variable;
-
-        /**
-         * Gets the value of the variable property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link Query.Variables.Variable }
-         *     
-         */
-        public Query.Variables.Variable getVariable() {
-            return variable;
-        }
-
-        /**
-         * Sets the value of the variable property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link Query.Variables.Variable }
-         *     
-         */
-        public void setVariable(Query.Variables.Variable value) {
-            this.variable = value;
-        }
-
-
-        /**
-         * <p>Java class for anonymous complex type.
-         * 
-         * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
-         * <pre>
-         * &lt;complexType&gt;
-         *   &lt;complexContent&gt;
-         *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-         *       &lt;sequence&gt;
-         *         &lt;element name="qname"&gt;
-         *           &lt;complexType&gt;
-         *             &lt;complexContent&gt;
-         *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-         *                 &lt;sequence&gt;
-         *                   &lt;element name="localname" type="{http://www.w3.org/2001/XMLSchema}NCName"/&gt;
-         *                   &lt;element name="namespace" type="{http://www.w3.org/2001/XMLSchema}anyURI"/&gt;
-         *                   &lt;element name="prefix" type="{http://www.w3.org/2001/XMLSchema}NCName" minOccurs="0"/&gt;
-         *                   &lt;element ref="{http://exist-db.org/xquery/types/serialized}sequence"/&gt;
-         *                 &lt;/sequence&gt;
-         *               &lt;/restriction&gt;
-         *             &lt;/complexContent&gt;
-         *           &lt;/complexType&gt;
-         *         &lt;/element&gt;
-         *       &lt;/sequence&gt;
-         *     &lt;/restriction&gt;
-         *   &lt;/complexContent&gt;
-         * &lt;/complexType&gt;
-         * </pre>
-         * 
-         * 
-         */
-        @XmlAccessorType(XmlAccessType.FIELD)
-        @XmlType(name = "", propOrder = {
-            "qname"
-        })
-        public static class Variable {
-
-            @XmlElement(required = true)
-            protected Query.Variables.Variable.Qname qname;
-
-            /**
-             * Gets the value of the qname property.
-             * 
-             * @return
-             *     possible object is
-             *     {@link Query.Variables.Variable.Qname }
-             *     
-             */
-            public Query.Variables.Variable.Qname getQname() {
-                return qname;
-            }
-
-            /**
-             * Sets the value of the qname property.
-             * 
-             * @param value
-             *     allowed object is
-             *     {@link Query.Variables.Variable.Qname }
-             *     
-             */
-            public void setQname(Query.Variables.Variable.Qname value) {
-                this.qname = value;
-            }
-
-
-            /**
-             * <p>Java class for anonymous complex type.
-             * 
-             * <p>The following schema fragment specifies the expected content contained within this class.
-             * 
-             * <pre>
-             * &lt;complexType&gt;
-             *   &lt;complexContent&gt;
-             *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-             *       &lt;sequence&gt;
-             *         &lt;element name="localname" type="{http://www.w3.org/2001/XMLSchema}NCName"/&gt;
-             *         &lt;element name="namespace" type="{http://www.w3.org/2001/XMLSchema}anyURI"/&gt;
-             *         &lt;element name="prefix" type="{http://www.w3.org/2001/XMLSchema}NCName" minOccurs="0"/&gt;
-             *         &lt;element ref="{http://exist-db.org/xquery/types/serialized}sequence"/&gt;
-             *       &lt;/sequence&gt;
-             *     &lt;/restriction&gt;
-             *   &lt;/complexContent&gt;
-             * &lt;/complexType&gt;
-             * </pre>
-             * 
-             * 
-             */
-            @XmlAccessorType(XmlAccessType.FIELD)
-            @XmlType(name = "", propOrder = {
-                "localname",
-                "namespace",
-                "prefix",
-                "sequence"
-            })
-            public static class Qname {
-
-                @XmlElement(required = true)
-                @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-                @XmlSchemaType(name = "NCName")
-                protected String localname;
-                @XmlElement(required = true)
-                @XmlSchemaType(name = "anyURI")
-                protected String namespace;
-                @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-                @XmlSchemaType(name = "NCName")
-                protected String prefix;
-                @XmlElement(namespace = "http://exist-db.org/xquery/types/serialized", required = true)
-                protected Sequence sequence;
-
-                /**
-                 * Gets the value of the localname property.
-                 * 
-                 * @return
-                 *     possible object is
-                 *     {@link String }
-                 *     
-                 */
-                public String getLocalname() {
-                    return localname;
-                }
-
-                /**
-                 * Sets the value of the localname property.
-                 * 
-                 * @param value
-                 *     allowed object is
-                 *     {@link String }
-                 *     
-                 */
-                public void setLocalname(String value) {
-                    this.localname = value;
-                }
-
-                /**
-                 * Gets the value of the namespace property.
-                 * 
-                 * @return
-                 *     possible object is
-                 *     {@link String }
-                 *     
-                 */
-                public String getNamespace() {
-                    return namespace;
-                }
-
-                /**
-                 * Sets the value of the namespace property.
-                 * 
-                 * @param value
-                 *     allowed object is
-                 *     {@link String }
-                 *     
-                 */
-                public void setNamespace(String value) {
-                    this.namespace = value;
-                }
-
-                /**
-                 * Gets the value of the prefix property.
-                 * 
-                 * @return
-                 *     possible object is
-                 *     {@link String }
-                 *     
-                 */
-                public String getPrefix() {
-                    return prefix;
-                }
-
-                /**
-                 * Sets the value of the prefix property.
-                 * 
-                 * @param value
-                 *     allowed object is
-                 *     {@link String }
-                 *     
-                 */
-                public void setPrefix(String value) {
-                    this.prefix = value;
-                }
-
-                /**
-                 * Gets the value of the sequence property.
-                 * 
-                 * @return
-                 *     possible object is
-                 *     {@link Sequence }
-                 *     
-                 */
-                public Sequence getSequence() {
-                    return sequence;
-                }
-
-                /**
-                 * Sets the value of the sequence property.
-                 * 
-                 * @param value
-                 *     allowed object is
-                 *     {@link Sequence }
-                 *     
-                 */
-                public void setSequence(Sequence value) {
-                    this.sequence = value;
-                }
-
-            }
-
-        }
-
-    }
-
-}
