--- exist-core/src/main/java/org/exist/http/jaxb/Result.java
@@ -1,353 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v3.0.0 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.12.01 at 10:16:19 AM CET 
-//
-
-/*
- * eXist-db Open Source Native XML Database
- * Copyright (C) 2001 The eXist-db Authors
- *
- * info@exist-db.org
- * http://www.exist-db.org
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-package org.exist.http.jaxb;
-
-import jakarta.xml.bind.annotation.*;
-import org.w3c.dom.Element;
-
-import java.util.ArrayList;
-import java.util.List;
-
-
-/**
- * <p>Java class for anonymous complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType&gt;
- *   &lt;complexContent&gt;
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *       &lt;choice&gt;
- *         &lt;element ref="{http://exist.sourceforge.net/NS/exist}collection"/&gt;
- *         &lt;element name="value" maxOccurs="unbounded" minOccurs="0"&gt;
- *           &lt;complexType&gt;
- *             &lt;complexContent&gt;
- *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *                 &lt;sequence&gt;
- *                   &lt;any processContents='skip'/&gt;
- *                 &lt;/sequence&gt;
- *                 &lt;attribute name="type" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
- *               &lt;/restriction&gt;
- *             &lt;/complexContent&gt;
- *           &lt;/complexType&gt;
- *         &lt;/element&gt;
- *       &lt;/choice&gt;
- *       &lt;attGroup ref="{http://exist.sourceforge.net/NS/exist}queryAttrs"/&gt;
- *     &lt;/restriction&gt;
- *   &lt;/complexContent&gt;
- * &lt;/complexType&gt;
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = {
-    "collection",
-    "value"
-})
-@XmlRootElement(name = "result")
-public class Result {
-
-    protected Collection collection;
-    protected List<Result.Value> value;
-    @XmlAttribute(name = "hits", namespace = "http://exist.sourceforge.net/NS/exist")
-    protected Integer hits;
-    @XmlAttribute(name = "start", namespace = "http://exist.sourceforge.net/NS/exist")
-    protected Integer start;
-    @XmlAttribute(name = "count", namespace = "http://exist.sourceforge.net/NS/exist")
-    protected Integer count;
-    @XmlAttribute(name = "compilation-time", namespace = "http://exist.sourceforge.net/NS/exist")
-    protected Integer compilationTime;
-    @XmlAttribute(name = "execution-time", namespace = "http://exist.sourceforge.net/NS/exist")
-    protected Integer executionTime;
-
-    /**
-     * Gets the value of the collection property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Collection }
-     *     
-     */
-    public Collection getCollection() {
-        return collection;
-    }
-
-    /**
-     * Sets the value of the collection property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Collection }
-     *     
-     */
-    public void setCollection(Collection value) {
-        this.collection = value;
-    }
-
-    /**
-     * Gets the value of the value property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the Jakarta XML Binding object.
-     * This is why there is not a <CODE>set</CODE> method for the value property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getValue().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link Result.Value }
-     * 
-     * @return the result value.
-     */
-    public List<Result.Value> getValue() {
-        if (value == null) {
-            value = new ArrayList<Result.Value>();
-        }
-        return this.value;
-    }
-
-    /**
-     * Gets the value of the hits property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getHits() {
-        return hits;
-    }
-
-    /**
-     * Sets the value of the hits property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setHits(Integer value) {
-        this.hits = value;
-    }
-
-    /**
-     * Gets the value of the start property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getStart() {
-        return start;
-    }
-
-    /**
-     * Sets the value of the start property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setStart(Integer value) {
-        this.start = value;
-    }
-
-    /**
-     * Gets the value of the count property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getCount() {
-        return count;
-    }
-
-    /**
-     * Sets the value of the count property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setCount(Integer value) {
-        this.count = value;
-    }
-
-    /**
-     * Gets the value of the compilationTime property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getCompilationTime() {
-        return compilationTime;
-    }
-
-    /**
-     * Sets the value of the compilationTime property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setCompilationTime(Integer value) {
-        this.compilationTime = value;
-    }
-
-    /**
-     * Gets the value of the executionTime property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getExecutionTime() {
-        return executionTime;
-    }
-
-    /**
-     * Sets the value of the executionTime property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setExecutionTime(Integer value) {
-        this.executionTime = value;
-    }
-
-
-    /**
-     * <p>Java class for anonymous complex type.
-     * 
-     * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
-     * <pre>
-     * &lt;complexType&gt;
-     *   &lt;complexContent&gt;
-     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
-     *       &lt;sequence&gt;
-     *         &lt;any processContents='skip'/&gt;
-     *       &lt;/sequence&gt;
-     *       &lt;attribute name="type" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
-     *     &lt;/restriction&gt;
-     *   &lt;/complexContent&gt;
-     * &lt;/complexType&gt;
-     * </pre>
-     * 
-     * 
-     */
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {
-        "content"
-    })
-    public static class Value {
-
-        @XmlMixed
-        @XmlAnyElement
-        protected List<Object> content;
-        @XmlAttribute(name = "type", namespace = "http://exist.sourceforge.net/NS/exist")
-        protected String type;
-
-        /**
-         * Gets the value of the content property.
-         * 
-         * <p>
-         * This accessor method returns a reference to the live list,
-         * not a snapshot. Therefore any modification you make to the
-         * returned list will be present inside the Jakarta XML Binding object.
-         * This is why there is not a <CODE>set</CODE> method for the content property.
-         * 
-         * <p>
-         * For example, to add a new item, do as follows:
-         * <pre>
-         *    getContent().add(newItem);
-         * </pre>
-         * 
-         * 
-         * <p>
-         * Objects of the following type(s) are allowed in the list
-         * {@link Element }
-         * {@link String }
-         * 
-         * @return the content.
-         */
-        public List<Object> getContent() {
-            if (content == null) {
-                content = new ArrayList<Object>();
-            }
-            return this.content;
-        }
-
-        /**
-         * Gets the value of the type property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link String }
-         *     
-         */
-        public String getType() {
-            return type;
-        }
-
-        /**
-         * Sets the value of the type property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link String }
-         *     
-         */
-        public void setType(String value) {
-            this.type = value;
-        }
-
-    }
-
-}
