--- exist-core/src/main/java/org/exist/http/jaxb/YesNo.java
@@ -1,77 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v3.0.0 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.12.01 at 10:16:19 AM CET 
-//
-
-/*
- * eXist-db Open Source Native XML Database
- * Copyright (C) 2001 The eXist-db Authors
- *
- * info@exist-db.org
- * http://www.exist-db.org
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-package org.exist.http.jaxb;
-
-import jakarta.xml.bind.annotation.XmlEnum;
-import jakarta.xml.bind.annotation.XmlEnumValue;
-import jakarta.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for yesNo.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * <pre>
- * &lt;simpleType name="yesNo"&gt;
- *   &lt;restriction base="{http://www.w3.org/2001/XMLSchema}NCName"&gt;
- *     &lt;enumeration value="yes"/&gt;
- *     &lt;enumeration value="no"/&gt;
- *   &lt;/restriction&gt;
- * &lt;/simpleType&gt;
- * </pre>
- * 
- */
-@XmlType(name = "yesNo")
-@XmlEnum
-public enum YesNo {
-
-    @XmlEnumValue("yes")
-    YES("yes"),
-    @XmlEnumValue("no")
-    NO("no");
-    private final String value;
-
-    YesNo(String v) {
-        value = v;
-    }
-
-    public String value() {
-        return value;
-    }
-
-    public static YesNo fromValue(String v) {
-        for (YesNo c: YesNo.values()) {
-            if (c.value.equals(v)) {
-                return c;
-            }
-        }
-        throw new IllegalArgumentException(v);
-    }
-
-}
