+++ exist-core/src/main/xjb/rest-api.xjb
@@ -1,6 +1,30 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
+    Elemental
+    Copyright (C) 2024, Evolved Binary Ltd
+
+    admin@evolvedbinary.com
+    https://www.evolvedbinary.com | https://www.elemental.xyz
+
+    This library is free software; you can redistribute it and/or
+    modify it under the terms of the GNU Lesser General Public
+    License as published by the Free Software Foundation; version 2.1.
+
+    This library is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    Lesser General Public License for more details.
+
+    You should have received a copy of the GNU Lesser General Public
+    License along with this library; if not, write to the Free Software
+    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+
+    NOTE: Parts of this file contain code from 'The eXist-db Authors'.
+          The original license header is included below.
+
+    =====================================================================
+
     eXist-db Open Source Native XML Database
     Copyright (C) 2001 The eXist-db Authors
 
@@ -25,12 +49,12 @@
 <jxb:bindings version="3.0" xmlns:jxb="https://jakarta.ee/xml/ns/jaxb" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
     <jxb:bindings schemaLocation="../xsd/rest-api.xsd" node="//xsd:schema">
         <jxb:schemaBindings>
-            <jxb:package name="org.exist.http.jaxb"/>
+            <jxb:package name="org.exist.http.rest"/>
         </jxb:schemaBindings>
     </jxb:bindings>
     <jxb:bindings schemaLocation="../xsd/rest-serialized-sequence.xsd" node="//xsd:schema">
         <jxb:schemaBindings>
-            <jxb:package name="org.exist.http.jaxb"/>
+            <jxb:package name="org.exist.http.rest.serialized"/>
         </jxb:schemaBindings>
     </jxb:bindings>
 </jxb:bindings>
